requirements.txt
python-telegram-bot==20.4
urllib3==1.26.5